refactor: 가용자금 컴포넌트 로직 리팩토링
